moved all image testing code to one module added min_corr image check; enabled by default to produce same results as assert_image_equal
