+++ vispy/testing/_testing.py
@@ -262,106 +262,6 @@ def requires_scipy(min_version='0.13'):
                                  'Requires Scipy version >= %s' % min_version)
 
 
-def _save_failed_test(data, expect, filename):
-    from ..io import _make_png
-    commit, error = run_subprocess(['git', 'rev-parse',  'HEAD'])
-    name = filename.split('/')
-    name.insert(-1, commit.strip())
-    filename = '/'.join(name)
-    host = 'data.vispy.org'
-
-    # concatenate data, expect, and diff into a single image
-    ds = data.shape
-    es = expect.shape
-    
-    shape = (max(ds[0], es[0]) + 4, ds[1] + es[1] + 8 + max(ds[1], es[1]), 4)
-    img = np.empty(shape, dtype=np.ubyte)
-    img[..., :3] = 100
-    img[..., 3] = 255
-    
-    img[2:2+ds[0], 2:2+ds[1], :ds[2]] = data
-    img[2:2+es[0], ds[1]+4:ds[1]+4+es[1], :es[2]] = expect
-    
-    diff = make_diff_image(data, expect)
-    img[2:2+diff.shape[0], -diff.shape[1]-2:-2] = diff
-
-    png = _make_png(img)
-    conn = httplib.HTTPConnection(host)
-    req = urllib.urlencode({'name': filename,
-                            'data': base64.b64encode(png)})
-    conn.request('POST', '/upload.py', req)
-    response = conn.getresponse().read()
-    conn.close()
-    print("\nImage comparison failed. Test result: %s %s   Expected result: "
-          "%s %s" % (data.shape, data.dtype, expect.shape, expect.dtype))
-    print("Uploaded to: \nhttp://%s/data/%s" % (host, filename))
-    if not response.startswith(b'OK'):
-        print("WARNING: Error uploading data to %s" % host)
-        print(response)
-
-
-def make_diff_image(im1, im2):
-    """Return image array showing the differences between im1 and im2.
-    
-    Handles images of different shape. Alpha channels are not compared.
-    """
-    ds = im1.shape
-    es = im2.shape
-    
-    diff = np.empty((max(ds[0], es[0]), max(ds[1], es[1]), 4), dtype=int)
-    diff[..., :3] = 128
-    diff[..., 3] = 255
-    diff[:ds[0], :ds[1], :min(ds[2], 3)] += im1[..., :3]
-    diff[:es[0], :es[1], :min(es[2], 3)] -= im2[..., :3]
-    diff = np.clip(diff, 0, 255).astype(np.ubyte)
-    return diff
-
-
-def assert_image_equal(image, reference, limit=0.9):
-    """Downloads reference image and compares with image
-
-    Parameters
-    ----------
-    image: str, numpy.array
-        'screenshot' or image data
-    reference: str
-        The filename on the remote ``test-data`` repository to download.
-    limit : float
-        The minimum acceptable cross-correlation value.
-    """
-    from ..gloo.util import _screenshot
-    from ..io import read_png
-    from ..geometry import resize
-
-    if image == "screenshot":
-        image = _screenshot(alpha=False)
-    if isinstance(reference, string_types):
-        ref = read_png(get_testing_file(reference))[:, :, :3]
-    else:
-        ref = reference
-        reference = 'ndarray'
-    assert isinstance(ref, np.ndarray) and ref.ndim == 3
-    # resize in case we're on a HiDPI display
-    image = resize(image, ref.shape[:2], 'nearest')
-
-    # check for minimum number of changed pixels, allowing for overall 1-pixel
-    # shift in any direcion
-    slice_as = [slice(0, -1), slice(0, None), slice(1, None)]
-    slice_bs = slice_as[::-1]
-    max_corr = -1
-    for ii in range(len(slice_as)):
-        for jj in range(len(slice_as)):
-            a = image[slice_as[ii], slice_as[jj]]
-            b = ref[slice_bs[ii], slice_bs[jj]]
-            with np.errstate(invalid='ignore'):
-                corr = np.corrcoef(a.ravel(), b.ravel())[0, 1]
-            if corr > max_corr:
-                max_corr = corr
-    if max_corr < limit:
-        _save_failed_test(image, ref, reference)
-        raise AssertionError('max_corr %s < %s' % (max_corr, limit))
-
-
 @nottest
 def TestingCanvas(bgcolor='black', size=(100, 100), dpi=None):
     """Class wrapper to avoid importing scene until necessary"""
