+++ vispy/visuals/tests/test_ellipse.py
@@ -32,7 +32,8 @@ def test_circle_draw():
         c.draw_visual(ellipse)
         # low corr here because borders have some variability
         # esp. w/HiDPI
-        assert_image_approved("screenshot", 'visuals/circle3.png')
+        assert_image_approved("screenshot", 'visuals/circle3.png', 
+                              min_corr=0.7)
 
 
 @requires_application()
@@ -59,7 +60,8 @@ def test_ellipse_draw():
         ellipse.transform = transforms.STTransform(scale=(2.0, 3.0),
                                                    translate=(50, 50))
         c.draw_visual(ellipse)
-        assert_image_approved("screenshot", 'visuals/ellipse3.png')
+        assert_image_approved("screenshot", 'visuals/ellipse3.png', 
+                              min_corr=0.7)
 
 
 @requires_application()
@@ -76,7 +78,8 @@ def test_arc_draw1():
                                   start_angle=150., span_angle=120.,
                                   border_color=(1, 0, 0, 1))
         c.draw_visual(ellipse)
-        assert_image_approved("screenshot", 'visuals/arc2.png')
+        assert_image_approved("screenshot", 'visuals/arc2.png', 
+                              min_corr=0.6)
 
 
 @requires_application()
